[backend] complete create products + delete products

+++ backend/internal/middleware/auth/handler.go
@@ -2,8 +2,8 @@ package auth
 
 import "net/http"
 
-func GetUserID(r *http.Request) (int64, bool) {
-	userID, ok := r.Context().Value(UserIDKey).(int64)
+func GetUserID(r *http.Request) (int32, bool) {
+	userID, ok := r.Context().Value(UserIDKey).(int32)
 	return userID, ok
 }
